move loadContract to util dir
